Rename module pas to pasplugin
